Module webidl::ast

Contains all structures related to the AST for the WebIDL grammar.

Structs

Argument

ArgumentListExtendedAttribute

Callback

CallbackInterface

Const

DictionaryMember

Enum

ExplicitStringifierOperation

IdentifierExtendedAttribute

IdentifierListExtendedAttribute

Implements
Note that this is no longer a part of the WebIDL specification, but exists for backwards compatibility of older WebIDLs.

ImplicitStringifierOperation

Includes

Iterable

Maplike

NamedArgumentListExtendedAttribute

NonPartialDictionary

NonPartialInterface

NonPartialMixin

NonPartialNamespace

PartialDictionary

PartialInterface

PartialMixin

PartialNamespace

RegularAttribute

RegularOperation

Setlike

SpecialOperation

StaticAttribute

StaticOperation

StringifierAttribute

Type

Typedef

Enums

Attribute

BufferRelatedType
Do not use - for grammar use only.

ConstType

ConstValue

DefaultValue

Definition

Dictionary

ExtendedAttribute

Interface

InterfaceMember

Mixin

MixinMember

Namespace

NamespaceMember

Operation

Other

OtherExtendedAttribute

OtherExtendedAttributeGroupType

PrimitiveType
Do not use - for grammar use only.

ReturnType

Special

StringType

Stringifier
Do not use - for grammar use only.

StringifierOperation

TypeKind

UnrestrictedFloatType
Do not use - for grammar use only.

UnsignedIntegerType
Do not use - for grammar use only.

Functions

flatten_asts
Consumes a vector of ASTs that are flattened into a single AST. This is helpful if you want to merge ASTs from multiple files and be able to use the visitor pattern across all of them.

Type Definitions

AST

Identifier