Trait webidl::visitor::ImmutableVisitor

pub trait ImmutableVisitor<'ast> {
    fn visit(&mut self, definitions: &'ast [Definition]) { ... }

    fn visit_argument(&mut self, argument: &'ast Argument) { ... }

    fn visit_argument_list_extended_attribute(
        &mut self, 
        ex: &'ast ArgumentListExtendedAttribute
    ) { ... }

    fn visit_attribute(&mut self, attribute: &'ast Attribute) { ... }

    fn visit_callback(&mut self, callback: &'ast Callback) { ... }

    fn visit_callback_interface(
        &mut self, 
        callback_interface: &'ast CallbackInterface
    ) { ... }

    fn visit_const(&mut self, const_: &'ast Const) { ... }

    fn visit_const_type(&mut self, const_type: &'ast ConstType) { ... }

    fn visit_const_value(&mut self, _const_value: &'ast ConstValue) { ... }

    fn visit_default_value(&mut self, default_value: &'ast DefaultValue) { ... }

    fn visit_definition(&mut self, definition: &'ast Definition) { ... }

    fn visit_dictionary(&mut self, dictionary: &'ast Dictionary) { ... }

    fn visit_dictionary_member(
        &mut self, 
        dictionary_member: &'ast DictionaryMember
    ) { ... }

    fn visit_enum(&mut self, enum_: &'ast Enum) { ... }

    fn visit_explicit_stringifier_operation(
        &mut self, 
        op: &'ast ExplicitStringifierOperation
    ) { ... }

    fn visit_extended_attribute(&mut self, ex: &'ast ExtendedAttribute) { ... }

    fn visit_identifier(&mut self, _identifier: &'ast str) { ... }

    fn visit_identifier_extended_attribute(
        &mut self, 
        ex: &'ast IdentifierExtendedAttribute
    ) { ... }

    fn visit_identifier_list_extended_attribute(
        &mut self, 
        ex: &'ast IdentifierListExtendedAttribute
    ) { ... }

    fn visit_implicit_stringifier_operation(
        &mut self, 
        op: &'ast ImplicitStringifierOperation
    ) { ... }

    fn visit_implements(&mut self, implements: &'ast Implements) { ... }

    fn visit_includes(&mut self, includes: &'ast Includes) { ... }

    fn visit_interface(&mut self, interface: &'ast Interface) { ... }

    fn visit_interface_member(
        &mut self, 
        interface_member: &'ast InterfaceMember
    ) { ... }

    fn visit_iterable(&mut self, iterable: &'ast Iterable) { ... }

    fn visit_maplike(&mut self, maplike: &'ast Maplike) { ... }

    fn visit_mixin(&mut self, mixin: &'ast Mixin) { ... }

    fn visit_mixin_member(&mut self, mixin_member: &'ast MixinMember) { ... }

    fn visit_named_argument_list_extended_attribute(
        &mut self, 
        ex: &'ast NamedArgumentListExtendedAttribute
    ) { ... }

    fn visit_namespace(&mut self, namespace: &'ast Namespace) { ... }

    fn visit_namespace_member(
        &mut self, 
        namespace_member: &'ast NamespaceMember
    ) { ... }

    fn visit_non_partial_dictionary(
        &mut self, 
        dictionary: &'ast NonPartialDictionary
    ) { ... }

    fn visit_non_partial_interface(
        &mut self, 
        interface: &'ast NonPartialInterface
    ) { ... }

    fn visit_non_partial_mixin(&mut self, mixin: &'ast NonPartialMixin) { ... }

    fn visit_non_partial_namespace(
        &mut self, 
        namespace: &'ast NonPartialNamespace
    ) { ... }

    fn visit_operation(&mut self, operation: &'ast Operation) { ... }

    fn visit_other(&mut self, other: &'ast Other) { ... }

    fn visit_other_extended_attribute(
        &mut self, 
        ex: &'ast OtherExtendedAttribute
    ) { ... }

    fn visit_partial_dictionary(&mut self, dictionary: &'ast PartialDictionary) { ... }

    fn visit_partial_interface(&mut self, interface: &'ast PartialInterface) { ... }

    fn visit_partial_mixin(&mut self, mixin: &'ast PartialMixin) { ... }

    fn visit_partial_namespace(&mut self, namespace: &'ast PartialNamespace) { ... }

    fn visit_regular_attribute(
        &mut self, 
        regular_attribute: &'ast RegularAttribute
    ) { ... }

    fn visit_regular_operation(
        &mut self, 
        regular_operation: &'ast RegularOperation
    ) { ... }

    fn visit_return_type(&mut self, return_type: &'ast ReturnType) { ... }

    fn visit_setlike(&mut self, setlike: &'ast Setlike) { ... }

    fn visit_special(&mut self, _special: &'ast Special) { ... }

    fn visit_special_operation(
        &mut self, 
        special_operation: &'ast SpecialOperation
    ) { ... }

    fn visit_static_attribute(
        &mut self, 
        static_attribute: &'ast StaticAttribute
    ) { ... }

    fn visit_static_operation(
        &mut self, 
        static_operation: &'ast StaticOperation
    ) { ... }

    fn visit_string_type(&mut self, _string_type: &'ast StringType) { ... }

    fn visit_stringifier_attribute(
        &mut self, 
        attribute: &'ast StringifierAttribute
    ) { ... }

    fn visit_stringifier_operation(
        &mut self, 
        operation: &'ast StringifierOperation
    ) { ... }

    fn visit_type(&mut self, type_: &'ast Type) { ... }

    fn visit_type_kind(&mut self, type_kind: &'ast TypeKind) { ... }

    fn visit_typedef(&mut self, typedef: &'ast Typedef) { ... }

    fn walk_argument(&mut self, argument: &'ast Argument) { ... }

    fn walk_argument_list_extended_attribute(
        &mut self, 
        ex: &'ast ArgumentListExtendedAttribute
    ) { ... }

    fn walk_attribute(&mut self, attribute: &'ast Attribute) { ... }

    fn walk_callback(&mut self, callback: &'ast Callback) { ... }

    fn walk_callback_interface(
        &mut self, 
        callback_interface: &'ast CallbackInterface
    ) { ... }

    fn walk_const(&mut self, const_: &'ast Const) { ... }

    fn walk_const_type(&mut self, const_type: &'ast ConstType) { ... }

    fn walk_default_value(&mut self, default_value: &'ast DefaultValue) { ... }

    fn walk_definition(&mut self, definition: &'ast Definition) { ... }

    fn walk_dictionary(&mut self, dictionary: &'ast Dictionary) { ... }

    fn walk_dictionary_member(
        &mut self, 
        dictionary_member: &'ast DictionaryMember
    ) { ... }

    fn walk_enum(&mut self, enum_: &'ast Enum) { ... }

    fn walk_explicit_stringifier_operation(
        &mut self, 
        op: &'ast ExplicitStringifierOperation
    ) { ... }

    fn walk_extended_attribute(&mut self, ex: &'ast ExtendedAttribute) { ... }

    fn walk_identifier_extended_attribute(
        &mut self, 
        ex: &'ast IdentifierExtendedAttribute
    ) { ... }

    fn walk_identifier_list_extended_attribute(
        &mut self, 
        ex: &'ast IdentifierListExtendedAttribute
    ) { ... }

    fn walk_implicit_stringifier_operation(
        &mut self, 
        op: &'ast ImplicitStringifierOperation
    ) { ... }

    fn walk_implements(&mut self, implements: &'ast Implements) { ... }

    fn walk_includes(&mut self, includes: &'ast Includes) { ... }

    fn walk_interface(&mut self, interface: &'ast Interface) { ... }

    fn walk_interface_member(&mut self, interface_member: &'ast InterfaceMember) { ... }

    fn walk_iterable(&mut self, iterable: &'ast Iterable) { ... }

    fn walk_maplike(&mut self, maplike: &'ast Maplike) { ... }

    fn walk_mixin(&mut self, mixin: &'ast Mixin) { ... }

    fn walk_mixin_member(&mut self, mixin_member: &'ast MixinMember) { ... }

    fn walk_named_argument_list_extended_attribute(
        &mut self, 
        ex: &'ast NamedArgumentListExtendedAttribute
    ) { ... }

    fn walk_namespace(&mut self, namespace: &'ast Namespace) { ... }

    fn walk_namespace_member(&mut self, namespace_member: &'ast NamespaceMember) { ... }

    fn walk_non_partial_dictionary(
        &mut self, 
        dictionary: &'ast NonPartialDictionary
    ) { ... }

    fn walk_non_partial_interface(
        &mut self, 
        interface: &'ast NonPartialInterface
    ) { ... }

    fn walk_non_partial_mixin(&mut self, mixin: &'ast NonPartialMixin) { ... }

    fn walk_non_partial_namespace(
        &mut self, 
        namespace: &'ast NonPartialNamespace
    ) { ... }

    fn walk_operation(&mut self, operation: &'ast Operation) { ... }

    fn walk_other(&mut self, other: &'ast Other) { ... }

    fn walk_other_extended_attribute(
        &mut self, 
        ex: &'ast OtherExtendedAttribute
    ) { ... }

    fn walk_partial_dictionary(
        &mut self, 
        partial_dictionary: &'ast PartialDictionary
    ) { ... }

    fn walk_partial_interface(
        &mut self, 
        partial_interface: &'ast PartialInterface
    ) { ... }

    fn walk_partial_mixin(&mut self, partial_mixin: &'ast PartialMixin) { ... }

    fn walk_partial_namespace(
        &mut self, 
        partial_namespace: &'ast PartialNamespace
    ) { ... }

    fn walk_regular_attribute(
        &mut self, 
        regular_attribute: &'ast RegularAttribute
    ) { ... }

    fn walk_regular_operation(
        &mut self, 
        regular_operation: &'ast RegularOperation
    ) { ... }

    fn walk_return_type(&mut self, return_type: &'ast ReturnType) { ... }

    fn walk_setlike(&mut self, setlike: &'ast Setlike) { ... }

    fn walk_special_operation(
        &mut self, 
        special_operation: &'ast SpecialOperation
    ) { ... }

    fn walk_static_attribute(&mut self, static_attribute: &'ast StaticAttribute) { ... }

    fn walk_static_operation(&mut self, static_operation: &'ast StaticOperation) { ... }

    fn walk_stringifier_attribute(
        &mut self, 
        attribute: &'ast StringifierAttribute
    ) { ... }

    fn walk_stringifier_operation(
        &mut self, 
        stringifier_operation: &'ast StringifierOperation
    ) { ... }

    fn walk_type(&mut self, type_: &'ast Type) { ... }

    fn walk_type_kind(&mut self, type_kind: &'ast TypeKind) { ... }

    fn walk_typedef(&mut self, typedef: &'ast Typedef) { ... }
}

Provided Methods

fn visit(&mut self, definitions: &'ast [Definition])

This is the initial function used to start traversing the AST. By default, this will simply recursively walk down the AST without performing any meaningful action.

fn visit_argument(&mut self, argument: &'ast Argument)

fn visit_argument_list_extended_attribute(
&mut self,
ex: &'ast ArgumentListExtendedAttribute
)

fn visit_attribute(&mut self, attribute: &'ast Attribute)

fn visit_callback(&mut self, callback: &'ast Callback)

fn visit_callback_interface(
&mut self,
callback_interface: &'ast CallbackInterface
)

fn visit_const(&mut self, const_: &'ast Const)

fn visit_const_type(&mut self, const_type: &'ast ConstType)

fn visit_const_value(&mut self, _const_value: &'ast ConstValue)

fn visit_default_value(&mut self, default_value: &'ast DefaultValue)

fn visit_definition(&mut self, definition: &'ast Definition)

fn visit_dictionary(&mut self, dictionary: &'ast Dictionary)

fn visit_dictionary_member(&mut self, dictionary_member: &'ast DictionaryMember)

fn visit_enum(&mut self, enum_: &'ast Enum)

fn visit_explicit_stringifier_operation(
&mut self,
op: &'ast ExplicitStringifierOperation
)

fn visit_extended_attribute(&mut self, ex: &'ast ExtendedAttribute)

fn visit_identifier(&mut self, _identifier: &'ast str)

fn visit_identifier_extended_attribute(
&mut self,
ex: &'ast IdentifierExtendedAttribute
)

fn visit_identifier_list_extended_attribute(
&mut self,
ex: &'ast IdentifierListExtendedAttribute
)

fn visit_implicit_stringifier_operation(
&mut self,
op: &'ast ImplicitStringifierOperation
)

fn visit_implements(&mut self, implements: &'ast Implements)

fn visit_includes(&mut self, includes: &'ast Includes)

fn visit_interface(&mut self, interface: &'ast Interface)

fn visit_interface_member(&mut self, interface_member: &'ast InterfaceMember)

fn visit_iterable(&mut self, iterable: &'ast Iterable)

fn visit_maplike(&mut self, maplike: &'ast Maplike)

fn visit_mixin(&mut self, mixin: &'ast Mixin)

fn visit_mixin_member(&mut self, mixin_member: &'ast MixinMember)

fn visit_named_argument_list_extended_attribute(
&mut self,
ex: &'ast NamedArgumentListExtendedAttribute
)

fn visit_namespace(&mut self, namespace: &'ast Namespace)

fn visit_namespace_member(&mut self, namespace_member: &'ast NamespaceMember)

fn visit_non_partial_dictionary(
&mut self,
dictionary: &'ast NonPartialDictionary
)

fn visit_non_partial_interface(&mut self, interface: &'ast NonPartialInterface)

fn visit_non_partial_mixin(&mut self, mixin: &'ast NonPartialMixin)

fn visit_non_partial_namespace(&mut self, namespace: &'ast NonPartialNamespace)

fn visit_operation(&mut self, operation: &'ast Operation)

fn visit_other(&mut self, other: &'ast Other)

fn visit_other_extended_attribute(&mut self, ex: &'ast OtherExtendedAttribute)

fn visit_partial_dictionary(&mut self, dictionary: &'ast PartialDictionary)

fn visit_partial_interface(&mut self, interface: &'ast PartialInterface)

fn visit_partial_mixin(&mut self, mixin: &'ast PartialMixin)

fn visit_partial_namespace(&mut self, namespace: &'ast PartialNamespace)

fn visit_regular_attribute(&mut self, regular_attribute: &'ast RegularAttribute)

fn visit_regular_operation(&mut self, regular_operation: &'ast RegularOperation)

fn visit_return_type(&mut self, return_type: &'ast ReturnType)

fn visit_setlike(&mut self, setlike: &'ast Setlike)

fn visit_special(&mut self, _special: &'ast Special)

fn visit_special_operation(&mut self, special_operation: &'ast SpecialOperation)

fn visit_static_attribute(&mut self, static_attribute: &'ast StaticAttribute)

fn visit_static_operation(&mut self, static_operation: &'ast StaticOperation)

fn visit_string_type(&mut self, _string_type: &'ast StringType)

fn visit_stringifier_attribute(&mut self, attribute: &'ast StringifierAttribute)

fn visit_stringifier_operation(&mut self, operation: &'ast StringifierOperation)

fn visit_type(&mut self, type_: &'ast Type)

fn visit_type_kind(&mut self, type_kind: &'ast TypeKind)

fn visit_typedef(&mut self, typedef: &'ast Typedef)

fn walk_argument(&mut self, argument: &'ast Argument)

fn walk_argument_list_extended_attribute(
&mut self,
ex: &'ast ArgumentListExtendedAttribute
)

fn walk_attribute(&mut self, attribute: &'ast Attribute)

fn walk_callback(&mut self, callback: &'ast Callback)

fn walk_callback_interface(
&mut self,
callback_interface: &'ast CallbackInterface
)

fn walk_const(&mut self, const_: &'ast Const)

fn walk_const_type(&mut self, const_type: &'ast ConstType)

fn walk_default_value(&mut self, default_value: &'ast DefaultValue)

fn walk_definition(&mut self, definition: &'ast Definition)

fn walk_dictionary(&mut self, dictionary: &'ast Dictionary)

fn walk_dictionary_member(&mut self, dictionary_member: &'ast DictionaryMember)

fn walk_enum(&mut self, enum_: &'ast Enum)

fn walk_explicit_stringifier_operation(
&mut self,
op: &'ast ExplicitStringifierOperation
)

fn walk_extended_attribute(&mut self, ex: &'ast ExtendedAttribute)

fn walk_identifier_extended_attribute(
&mut self,
ex: &'ast IdentifierExtendedAttribute
)

fn walk_identifier_list_extended_attribute(
&mut self,
ex: &'ast IdentifierListExtendedAttribute
)

fn walk_implicit_stringifier_operation(
&mut self,
op: &'ast ImplicitStringifierOperation
)

fn walk_implements(&mut self, implements: &'ast Implements)

fn walk_includes(&mut self, includes: &'ast Includes)

fn walk_interface(&mut self, interface: &'ast Interface)

fn walk_interface_member(&mut self, interface_member: &'ast InterfaceMember)

fn walk_iterable(&mut self, iterable: &'ast Iterable)

fn walk_maplike(&mut self, maplike: &'ast Maplike)

fn walk_mixin(&mut self, mixin: &'ast Mixin)

fn walk_mixin_member(&mut self, mixin_member: &'ast MixinMember)

fn walk_named_argument_list_extended_attribute(
&mut self,
ex: &'ast NamedArgumentListExtendedAttribute
)

fn walk_namespace(&mut self, namespace: &'ast Namespace)

fn walk_namespace_member(&mut self, namespace_member: &'ast NamespaceMember)

fn walk_non_partial_dictionary(
&mut self,
dictionary: &'ast NonPartialDictionary
)

fn walk_non_partial_interface(&mut self, interface: &'ast NonPartialInterface)

fn walk_non_partial_mixin(&mut self, mixin: &'ast NonPartialMixin)

fn walk_non_partial_namespace(&mut self, namespace: &'ast NonPartialNamespace)

fn walk_operation(&mut self, operation: &'ast Operation)

fn walk_other(&mut self, other: &'ast Other)

fn walk_other_extended_attribute(&mut self, ex: &'ast OtherExtendedAttribute)

fn walk_partial_dictionary(
&mut self,
partial_dictionary: &'ast PartialDictionary
)

fn walk_partial_interface(&mut self, partial_interface: &'ast PartialInterface)

fn walk_partial_mixin(&mut self, partial_mixin: &'ast PartialMixin)

fn walk_partial_namespace(&mut self, partial_namespace: &'ast PartialNamespace)

fn walk_regular_attribute(&mut self, regular_attribute: &'ast RegularAttribute)

fn walk_regular_operation(&mut self, regular_operation: &'ast RegularOperation)

fn walk_return_type(&mut self, return_type: &'ast ReturnType)

fn walk_setlike(&mut self, setlike: &'ast Setlike)

fn walk_special_operation(&mut self, special_operation: &'ast SpecialOperation)

fn walk_static_attribute(&mut self, static_attribute: &'ast StaticAttribute)

fn walk_static_operation(&mut self, static_operation: &'ast StaticOperation)

fn walk_stringifier_attribute(&mut self, attribute: &'ast StringifierAttribute)

fn walk_stringifier_operation(
&mut self,
stringifier_operation: &'ast StringifierOperation
)

fn walk_type(&mut self, type_: &'ast Type)

fn walk_type_kind(&mut self, type_kind: &'ast TypeKind)

fn walk_typedef(&mut self, typedef: &'ast Typedef)